refactor(common)!: move timeline service client to org.apache.hudi.common.table.timeline
Relocates TimelineServiceClient and TimelineServiceClientBase (plus
test) from the two-class org.apache.hudi.timeline package into
org.apache.hudi.common.table.timeline, next to the timeline model and
the dto package whose payloads the client fetches. Removes the
org.apache.hudi.timeline package from hudi-common, disambiguating it
from the timeline-service server module namespace.

BREAKING CHANGE: source/binary break for external code importing the
client directly; unlikely surface - the client is constructed
internally (RemoteHoodieTableFileSystemView, timeline-server-based
markers), never named in configs or loaded reflectively. Fails loudly.
